Wilt simulation 30 Oct 2025
This is an initial model for the spread of Oak wilt on the Apostle islands. It should operate as a base without any parameters. All Cels with oaks present are susceptible and one cell is randomly infected. The simulation runs for x timesteps (50 right now) and infects all neighboring cells each year